+++ bluedragon-distribution-core/src/main/java/com/jd/bluedragon/distribution/station/service/impl/WorkStationServiceImpl.java
@@ -5,9 +5,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import com.jd.bluedragon.distribution.station.enums.BusinessLineTypeEnum;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
-import org.apache.lucene.util.CollectionUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,6 +19,7 @@ import com.jd.bluedragon.distribution.api.response.base.Result;
 import com.jd.bluedragon.distribution.station.dao.WorkStationDao;
 import com.jd.bluedragon.distribution.station.domain.DeleteRequest;
 import com.jd.bluedragon.distribution.station.domain.WorkStation;
+import com.jd.bluedragon.distribution.station.domain.WorkStationAttendPlan;
 import com.jd.bluedragon.distribution.station.domain.WorkStationCountVo;
 import com.jd.bluedragon.distribution.station.query.WorkStationQuery;
 import com.jd.bluedragon.distribution.station.service.WorkStationGridService;
@@ -82,6 +83,7 @@ public class WorkStationServiceImpl implements WorkStationService {
 			}else {
 				generateAndSetBusinessKey(data);
 			}
+			data.setBusinessLineName(BusinessLineTypeEnum.getNameByCode(data.getBusinessLineCode()));
 			workStationDao.insert(data);
 		}
 		return result;
@@ -125,6 +127,9 @@ public class WorkStationServiceImpl implements WorkStationService {
 			if(uniqueKeysRowNumMap.containsKey(uniqueKeysStr)) {
 				return result0.toFail(rowKey + "和第"+uniqueKeysRowNumMap.get(uniqueKeysStr)+"行数据重复！");
 			}
+			if(BusinessLineTypeEnum.getEnum(data.getBusinessLineCode()) == null){
+				return result0.toFail(rowKey + "的【业务条线ID】不符合要求！");
+			}
 			uniqueKeysRowNumMap.put(uniqueKeysStr, rowNum);
 			rowNum ++;
 		}
@@ -149,6 +154,7 @@ public class WorkStationServiceImpl implements WorkStationService {
 		String workName = data.getWorkName();
 		String areaCode = data.getAreaCode();
 		String areaName = data.getAreaName();
+		data.setBusinessLineName(BusinessLineTypeEnum.getNameByCode(data.getBusinessLineCode()));
 		
 		if(!CheckHelper.checkStr("作业区ID", areaCode, 50, result).isSuccess()) {
 			return result;
@@ -174,8 +180,13 @@ public class WorkStationServiceImpl implements WorkStationService {
 		if(!result.isSuccess()) {
 			return result;
 		}
+		WorkStation oldData = workStationDao.queryById(updateData.getId());
+		if(oldData == null) {
+			return result.toFail("该工序数据已变更，请重新查询后修改！");
+		}
 		workStationDao.deleteById(updateData);
 		updateData.setId(null);
+		updateData.setBusinessLineName(BusinessLineTypeEnum.getNameByCode(updateData.getBusinessLineCode()));
 		result.setData(workStationDao.insert(updateData) == 1);
 		return result;
 	 }
